`TextBlot.split` shouldn't split text if `index` is invalid
This can happen if the `length` value created in Quill's `Editor.applyDelta`
method was created from a string containing a unicode character with multiple
code points. For example, this operation contains `ú` or `\u0075\u0301`:

```
{
  "insert": "LaCroix's Cúrate collection of sparkling waters",
  "attributes": {
    "link": "http://www.lacroixwater.com/"
   }
}
```

The problem arises because the computed length of that string in `applyDelta` is 48,
but later in the Quill lifecycle* the string length is 47 because the `ú` has been
converted to a `ú` or `\u00FA`. Thus, `TextBlot.split` is called with an `index` that
is greater than the length of the string, resulting in the following error:

```
DOMException: Failed to execute 'splitText' on 'Text': The offset 48 is larger than the Text node's length.
    at TextWithSoftWrapsBlot.e.split
```

* Tested in Chrome 70.0.3538.110 on MacOS 10.14.1