Fix compatibility with TypeScript 3.7
TypeScript 3.7 introduces some changes to DOM related types, that `addedNodes` is
now a `NodeList` instead of `Node[]`, so some code won't compile.

This pr changes `.apply` to spread syntax to address this issue. Performance-wise
thery are similar and compatibility-wise we've already used spread syntax in the
codebase so that's should be fine.

Closes #72, #78.